+++ src/declarations/index.d.ts
@@ -39,6 +39,7 @@ interface IGlobalCache {
 	numbers: { [key: string]: number };
 	lists: { [key: string]: any[] };
 	costMatrices: { [key: string]: CostMatrix };
+	roomPositions: { [key: string]: RoomPosition | undefined };
 	things: { [key: string]: undefined | HasID | HasID[] };
 	// objects: { [key: string]: Object };
 }